#1 - 타이머 코드 리펙토링
- 필요없는 파일 병합 및 삭제
- 타이머 앱 생명주기에 의한 변화 내용 주석으로 문서화
- 뷰 재사용을 위함 모귤화작업
- 타이머 모드 아이콘 이미지 추가
